% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterwithoutlier.R
\name{scatterwithoutlier}
\alias{scatterwithoutlier}
\title{Create scatterplot for two quantitative variables with labeled outliers}
\usage{
scatterwithoutlier(
  data,
  xy = TRUE,
  cooksd = 0.5,
  outputcount = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{a list (dataframe, tibble, etc…) with two quantitative variables.}

\item{xy}{ogicals. If TRUE then the first column of the list would assumed to be the independent variable, and vice versa for FALSE. The default if TRUE.}

\item{cooksd}{a non-negative numerical value for cook’s distance threshold. The default is 0.5.}

\item{outputcount}{logicals. If TRUE then print a message on the count of influential observations.}

\item{verbose}{logicals. If TRUE then print a message to inform the user that the process is running.}

\item{...}{additional arguments to be passed to the low level regression fitting functions.}
}
\value{
A ggplot object(scatterplot) with fitted regression line and red labelled influential points.
}
\description{
Create scatterplot for two quantitative variables with labeled outliers
}
\examples{

scatterwithoutlier(gapminder \%>\%
select(lifeExp,pop))

scatterwithoutlier(gapminder \%>\%
select(pop,lifeExp),xy=FALSE,cooksd=0.05,verbose = TRUE,outputcount = TRUE)

scatterwithoutlier(gapminder \%>\%
select(lifeExp,gdpPercap),cooksd = 0.05)
}
